requirements.txt
fastmcp>=2.0,<3
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
